+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
 - Expose delete app org group admin API [#312](https://github.com/rokwire/core-building-block/issues/312)
+
 ### Fixed
 - Fix create role and create group Admin APIs [#386](https://github.com/rokwire/core-building-block/issues/386)